feature: add endpoints configurations to send and get data.
