+++ src/controller/assistant.ts
@@ -1,7 +1,8 @@
 import { drizzle } from 'drizzle-orm/neon-http';
 import { neon } from '@neondatabase/serverless';
-import { assistant, song } from './db/schema';
+import { assistant } from '../db/schema';
 import { Hono } from 'hono';
+import { S } from 'vitest/dist/reporters-LqC_WI4d.js';
 
 export type Env = {
   DATABASE_URL: string;
@@ -9,8 +10,6 @@ export type Env = {
 
 const routes = new Hono<{ Bindings: Env }>();
 
-
-//------------ songs api -------------------
 
 routes.get('/assistant', async (c) => {
   try {
@@ -34,7 +33,7 @@ routes.get('/assistant', async (c) => {
 });
 
 routes.post('/assistant', async (c) => {
-    const body = await c.req.json()
+    const body = await c.req.parseBody()
     const new_assitant = {
         full_name: body.full_name as string,
         is_going: body.is_going as string,
@@ -83,79 +82,4 @@ routes.delete('/assistant:id', async (c) => {
     }
   });
 
-//------------ songs api -------------------
-
-routes.get('/song', async (c) => {
-  try {
-    const sql = neon(c.env.DATABASE_URL);
-    const db = drizzle(sql);
-
-    const songs = await db.select().from(song);
-
-    return c.json({
-      songs,
-    });
-  } catch (error) {
-    console.log(error);
-    return c.json(
-      {
-        error,
-      },
-      400
-    );
-  }
-});
-
-routes.post('/song', async (c) => {
-  const body = await c.req.json()
-  
-  console.log('Data recibida:', body);
-  
-  const new_song = {
-      guest_name: body.guest_name as string,
-      song_name: body.song_name as string,
-      url: body.url as string
-  }
-  
-  console.log('Data a enviar a la DB:', new_song);
-  try {
-    const sql = neon(c.env.DATABASE_URL);
-    const db = drizzle(sql);
-
-    await db.insert(song).values(new_song);
-
-    return new Response('Created', { status: 201 })
-  } catch (error) {
-    console.log(error);
-    return c.json(
-      {
-        error,
-      },
-      400
-    );
-  }
-});
-
-
-routes.delete('/song:id', async (c) => {
-  const userId = await c.req.param('id')
-
-  try {
-    const sql = neon(c.env.DATABASE_URL);
-    const db = drizzle(sql);
-
-    await db.delete(song).where(eq(song.id as any, userId));
-
-    return new Response('Deleted', { status: 200 })
-  } catch (error) {
-    console.log(error);
-    return c.json(
-      {
-        error,
-      },
-      400
-    );
-  }
-});
-
-export default routes;
+export const assistantRoutes = routes
