+++ src/controller/index.ts
@@ -1,0 +1,2 @@
+export { assistantRoutes } from './assistant'
+export { songRoutes } from './song'
